# Conflicts:
#	en/images/right_bottom--old.png

+++ admin/js/custom/edit-item.js
@@ -39,19 +39,43 @@ $('#price').bind('input', function() {
 
 function delete_item(item_id,url)
 {
-    addLoading();
-    $.ajax({
-        url:"ajax/delete_item.php",
-        method:"post",
-        data:{item_id:item_id},
-        dataType:"json",
-        success:function(data)
-        {
-            hideLoading();
-            alert("Subitems deleted successfully");
-            window.location.href = url;
-        }
+    $(function(){
+        swal({
+                title: "Are you sure?",
+                text: "You will not be able to recover this Item!",
+                type: "warning",
+                showCancelButton: true,
+                confirmButtonColor: "#DD6B55",
+                confirmButtonText: "Yes, delete it!",
+                cancelButtonText: "No, cancel please!",
+                closeOnConfirm: false,
+                closeOnCancel: false
+            },
+            function(isConfirm){
+                if (isConfirm) {
+                    addLoading();
+                    $.ajax({
+                        url:"ajax/delete_item.php",
+                        method:"post",
+                        data:{item_id:item_id},
+                        dataType:"json",
+                        success:function(data)
+                        {
+                            hideLoading();
+                            alert("Item deleted successfully");
+                            window.location.href = url;
+                        }
+                    });
+                } else {
+                    swal("Cancelled", "Item is safe :)", "error");
+                    window.location.href = url;
+                }
+            });
     });
+
+
+
+
 }
 
 function edit_item(item_id,url)
@@ -100,14 +124,14 @@ function edit_item(item_id,url)
         'name_en'                 :  $('#name_en').val(),
         'name_he'                 :  $('#name_he').val(),
 
-        'desc_en'          :  $('#desc_en').val(),
-        'desc_he'          :  $('#desc_he').val(),
+        'desc_en'                 :  $('#desc_en').val(),
+        'desc_he'                 :  $('#desc_he').val(),
 
         'price'                   :  $('#price').val(),
 
         'hide'                    :  $('#hide').val(),
 
-        'item_id'             :   item_id
+        'item_id'                 :   item_id,
 
     };
 
@@ -120,7 +144,7 @@ function edit_item(item_id,url)
         success:function(data)
         {
             hideLoading();
-            alert("item added successfully");
+            alert("item updated successfully");
             window.location.href = url;
         }
     });
